fix localecompare bug
