+++ packages/garden/src/lib/hooks/index.ts
@@ -1,3 +1,8 @@
 export * from './useData';
 export * from './useExpand';
 export * from './useReactiveValue';
+export * from './useGardenGroups';
+export * from './useGroupingKeys';
+export * from './useItemWidths';
+export * from './useRefresh';
+export * from './useVirtualScrolling';
