+++ packages/garden/src/lib/hooks/useItemWidths.ts
@@ -1,0 +1,27 @@
+import { useEffect, useMemo, useState } from 'react';
+import { useGardenContext } from '../components';
+import { useGardenGroups } from './useGardenGroups';
+
+export function useItemWidths() {
+	const [widths, setWidths] = useState<number[]>([]);
+	const groups = useGardenGroups();
+
+	const amountOfColumns = useMemo(() => groups.length, [groups]);
+
+	const {
+		visuals: { calculateItemWidth },
+		grouping: {
+			value: { horizontalGroupingAccessor },
+		},
+		customGroupByKeys,
+	} = useGardenContext();
+
+	useEffect(() => {
+		if (groups && amountOfColumns > 0) {
+			const width = calculateItemWidth(groups, horizontalGroupingAccessor, customGroupByKeys);
+			setWidths(new Array(amountOfColumns).fill(width));
+		}
+	}, [amountOfColumns, groups, calculateItemWidth, customGroupByKeys, horizontalGroupingAccessor]);
+
+	return widths;
+}
