+++ packages/test-app/src/app/app.tsx
@@ -13,7 +13,17 @@ function Workspace() {
 				nodeLabelCallback: (s) => s.sequenceNumber.toString(),
 				objectIdentifier: 'id',
 				initialGrouping: { horizontalGroupingAccessor: 'id', verticalGroupingKeys: [] },
-				fieldSettings: {},
+				fieldSettings: {
+					title: {
+						getKey: (s) => s.title,
+						getColumnSort: (a, b) => {
+							return a.localeCompare(b) === 1 ? -1 : 1;
+						},
+					},
+					sequenceNumber: {
+						getKey: (s) => s.sequenceNumber as unknown as string,
+					},
+				},
 			})
 			.addMiddleware((mediator) => {
 				mediator.onClick((click) => {
